python/home-automation-pipelines/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "home-automation-pipelines"
description = 'Pipelines mostly written in Python for syncing up some stuff around the house'
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
keywords = []
authors = [
  { name = "Nicholas Payne", email = "github@pype.aleeas.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "typer",
  "prefect",
  "httpx",
  "python-dotenv",
  "fuzzywuzzy",
  "python-Levenshtein",
  "rich",
  "webdavclient3",
  "ipython",
  "pandas",
  "tqdm",
  "eyed3"  # id3 tags mp3
]
dynamic = ["version"]

[project.urls]
Documentation = "https://github.com/unknown/home-automation-pipelines#readme"
Issues = "https://github.com/unknown/home-automation-pipelines/issues"
Source = "https://github.com/unknown/home-automation-pipelines"

[project.scripts]
home-automation-pipelines = "home_automation_pipelines.cli:home_automation_pipelines"

[tool.hatch.version]
path = "home_automation_pipelines/__about__.py"

[tool.hatch.envs.default]
dependencies = [
  "pytest",
  "pytest-cov",
]
[tool.hatch.envs.default.scripts]
cov = "pytest --cov-report=term-missing --cov-config=pyproject.toml --cov=home_automation_pipelines --cov=tests {args}"
no-cov = "cov --no-cov {args}"

[[tool.hatch.envs.test.matrix]]
python = ["310", "311"]

[tool.coverage.run]
branch = true
parallel = true
omit = [
  "home_automation_pipelines/__about__.py",
]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
